% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.lgcm.R
\name{bivariateLgcm}
\alias{bivariateLgcm}
\title{Bivariate Latent Growth Curve Model}
\usage{
bivariateLgcm(lgcm.x, lgcm.y, slope.x.slope.y.covariance = 0,
  icept.x.slope.y.covariance = 0, icept.y.slope.x.covariance = 0,
  icept.x.icept.y.covariance = 0, latent.correlations = FALSE)
}
\arguments{
\item{lgcm.x}{LGCM object for first process (refered to as x)}

\item{lgcm.y}{LGCM object for second process (refered to as y)}

\item{slope.x.slope.y.covariance}{Latent covariance of slope of x and slope of y}

\item{icept.x.slope.y.covariance}{Latent covariance of intercept of x and slope of y}

\item{icept.y.slope.x.covariance}{Latent covariance of slope of x and intercept of y}

\item{icept.x.icept.y.covariance}{Latent covariance of intercept of x and intercept of y}

\item{latent.correlations}{Specify a latent correlation matrix instead of a covariance matrix.}
}
\description{
Bivariate Latent Growth Curve Model
}

