% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.ecr.R
\name{ecr.generic}
\alias{ecr.generic}
\title{generic Effective Curve Reliability (ECR)}
\usage{
ecr.generic(model, constraints = "slope.variance")
}
\arguments{
\item{model}{A semper model}

\item{constraints}{A list of parameter names assumed to be zero in the null hypothesis}
}
\description{
generic Effective Curve Reliability (ECR)
}

