% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcm.R
\name{ecr}
\alias{ecr}
\alias{effectiveError}
\alias{grr}
\title{Reliability and Precision indices}
\usage{
ecr(lgcm)
}
\arguments{
\item{lgcm}{A \code{lgcm} object representing a linear latent grwoth curve model}
}
\description{
A set of indicators for study design efficiency. Effective error, Effective Curve Reliability, and Growth Rate Reliability.
}

