% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcm.R
\name{effectiveError}
\alias{effectiveError}
\title{Compute effective error from a LGCM}
\usage{
effectiveError(lgcm)
}
\arguments{
\item{lgcm}{A \code{lgcm} object.}
}
\description{
Compute effective error from a LGCM
}

