\name{getOpenMxRepresentation}
\alias{getOpenMxRepresentation}


\title{Create an OpenMx model from an sempower model abstraction}
\description{This function creates a lgcm object to be run in OpenMx}

\usage{
getOpenMxRepresentation <- function(generic.model)		 
				 
}

\arguments{
	\item{generic.model}{A sempower model abstraction}
}


\author{Andreas M. Brandmaier}

