% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcm.R
\name{lgcm}
\alias{lgcm}
\title{Create a latent growth curve model}
\usage{
lgcm(timepoints = 0, intercept.variance = 0, slope.variance = 0,
  residual.variance = 0, intercept.slope.covariance = 0, sample.size = 0,
  intercept.mean = 0, slope.mean = 0)
}
\arguments{
\item{timepoints}{Vector of time points of measurement occasions}

\item{intercept.variance}{Individual differences in intercept}

\item{slope.variance}{Individual differences in slopes}

\item{residual.variance}{Residual variance or inverse of the precision of measurement}

\item{sample.size}{Number of cases}
}
\description{
This function creates a lgcm object representing a linear latent growth curve model
}

