% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePower.R
\name{power}
\alias{power}
\alias{sampleSize}
\title{Estimate Power or Sample Size from a simulation}
\usage{
power(psim, alpha = 0.05, exclude = c())
}
\arguments{
\item{psim}{A simPower result object.}

\item{alpha}{Alpha level for power estimation}

\item{power}{Target power}
}
\value{
Return either the sample size if power was set with \code{estimatePower} or the power if sample size was fixed with \code{estimateSampleSize}.
}
\description{
Estimate Power or Sample Size from a simulation
}
\examples{

power(psim, alpha=.05)
sampleSize(psim, power=0.8)


}
\author{
Andreas M. Brandmaier
}
\seealso{
\code{\link{simPower}}
}

