% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlcm.R
\name{qgcm}
\alias{qgcm}
\title{Quadratic Growth Curve Model}
\usage{
qgcm(timepoints = 0, intercept.variance = 0, slope.variance = 0,
  quadratic.variance = 0, residual.variance = 0,
  intercept.slope.covariance = 0, intercept.quadratic.covariance = 0,
  slope.quadratic.covariance = 0, sample.size = 0, intercept.mean = 0,
  slope.mean = 0, quadratic.mean = 0)
}
\description{
Quadratic Growth Curve Model
}

