\name{reliability}
\alias{ecr}
\alias{grr}
\alias{effective.error}

\title{Reliability and Precision indices}
\description{A set of indicators for study design efficiency. Effective error, Effective Curve Reliability, and Growth Rate Reliability.}

\usage{
effective.error(lgcm)
ecr(lgcm)
grr(lgcm)
}

\arguments{
	\item{lgcm}{A \code{lgcm} object representing a linear latent grwoth curve model}
}


\author{Andreas M. Brandmaier}

