% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPower.R
\name{simPower}
\alias{plot.simPower}
\alias{print.simPower}
\alias{simPower}
\title{Monte-Carlo power simulation}
\usage{
simPower(h0Model, h1Model, populationModel, N = c(20, 100),
  repetitions = 100, keepModels = FALSE, simfunc = simulateData,
  simargs = NULL)
}
\arguments{
\item{h0Model}{Model representing the null hypothesis}

\item{h1Model}{Model representing the alternative hypothesis}

\item{N}{Sample size. For RAM models, this is either (1) a scalar representing a fixed sample size, (2) a list with two items representing a minimum and a maximum for randomly drawn sample sizes (when sample size is the dependent variable), or (3) for multiple group models, this is a named list of sample sizes for the sub models where the names must match the submodels' names.}

\item{repetitions}{Number of Monte Carlo trials}

\item{keepModels}{Boolean. Keep models from all trials for later inspection?}

\item{simfunc}{Function to simulate data. Default is based on multivariate normal distribution.}

\item{simargs}{Extra arguments to be passed to the method for simulating data.}

\item{populationModel}{True}{population model used to generate data from.}

\item{psim}{SimPower result object}

\item{prng}{Range for plotting p values}

\item{target.power}{If not null, draw a line to indicate sample size for a specified target power in a power curve plot.}

\item{lw}{line width for power curve}

\item{lty}{line type for power curve}
}
\description{
Power simulation of a likelihood ratio test between two
models given simulated data from a true model in R/OpenMx
Assume you have a H1 model that test whether a pretest and a posttest score
}
\examples{
simPower(h0Model,h1Model, populationModel, N=c(20,100),repetitions=100, keepModels=F)

plot(psim, prng=seq(0.4,0.95,0.05), target.power=NULL, lw=2, lty=1,
                         xlab="sample size", ylab="statistical power",
                          main="Monte Carlo Power Simulation", add=F,...)
					   
print(psim, ...)	

}
\author{
: Andreas Brandmaier
}
\seealso{
\code{\link{simPowerZeroRestriction}}
}

