% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{Simulate Data}
\usage{
simulateData(model, N, format = "wide")
}
\arguments{
\item{model}{An OpenMx model.}

\item{N}{Desired sample size.}

\item{format}{Either "long" or "wide".}
}
\description{
Simulating data from a SEM by drawing from the multivariate normal as described by the model covariance matrix and means vector.
}
\examples{
simulateData(
toOpenMx(lgcm(slope.variance=1, 
intercept.variance=2, timepoints=0:8), N=100))

}
\author{
Andreas M. Brandmaier
}

